requirements.txt
wheel==0.36.2
Django==3.1.5
gunicorn==20.0.4
uritemplate==3.0.1
Markdown==3.3.3
Pygments==2.7.4
django-filter==2.4.0
django-guardian==2.3.0
djangorestframework==3.12.2
coreapi==2.3.3